Merged PR 5841: Updated the version number since we moved to console build to support NX
BASECORE recently moved to `CONSOLE` as subsystem during linking process, in order to support NX flag.

However, the side effect of this change is enforcing the subsystem level to be 5.0* at minimum, according to this document:
https://docs.microsoft.com/en-us/cpp/build/reference/subsystem-specify-subsystem?view=msvc-170#remarks

This change updated the subsystem level to be 6.02* to be higher than the default value.

Patch level is also updated to reflect that the security patches applied from last refresh.