Fixed missing length check in ConvertSmiHandlerSwContext() (#23)
ConvertSmiHandlerSwContext() is called by SmiHandlerProfileRegisterHandler() which is called by ProcessUserHandlerReg() when the supervisor dispatcher handles a SMM_MM_HDL_REG_1/SMM_MM_HDL_REG_2 request to register a child SMI handler with a gEfiSmmSwDispatch2ProtocolGuid handler guid. ConvertSmiHandlerSwContext() is handed a user space provided context and context size, and it will convert an EFI_SMM_SW_REGISTER_CONTEXT structure to a SMI_HANDLER_PROFILE_SW_REGISTER_CONTEXT structure. It doesn't check the context size and simply assumes it's large enough to hold the data it needs (it only does an assert). This could lead to user space inducing the supervisor to touch supervisor pages and possibly info leak some supervisor page data (a couple of bytes).

However, this is a debug feature for testing and development purposes only and should not be enabled in release products. Thus the security impact should be minor.

Fix:
- Turning the ASSERT length check into a proper length check (so that it would also trigger in release builds).

fixes #12