Fixed global length field updated even allocation in MmInstallConfigurationTable() failed (#18)
Bug Description:
MmInstallConfigurationTable() is called by the supervisor to allow user space to install configuration tables. It looks up the guid, and if it can't find it in the current table, it adds an entry. If there is no room left in the tables, they get reallocated to account for extra space. This reallocation has an interesting logic error. It updates the global variable (mMmSystemTableAllocateSize) that holds the table's max size prior to allocation. If the allocation fails (and hence no reallocation occurs), the global size variable isn't reset. This means the next caller that wants to add an entry would trigger memory corruption.

This has the potential for a compromised SMM handler to break into the supervisor, assuming a full memory corruption exploit. One of the prerequisites would be getting AllocatePool() to fail; however, I suspect this could be done utilizing a memory leak.

Fix:
This is fixed by using a temporary size value before allocation and assigning it to mMmSystemTableAllocateSize after successful allocation.

fixes #7