Fixing an edge case of incoming parameter check (#26)
This change fixed a potential overflow where if the maximal allowed number of pages being passed in (0x10000), the return will be an address of 0 pages.

The security impact is minimal as on allocation, the returned address will be either pointing to a free page that is marked with supervisor privilege or guarded page with read prevention protection. On the freeing, the input address will fail the internal attribute look up and bail early.