MmSupervisorPkg: MmIplX64Relay: Fix unintended return status
The MM core entry point address, as discovered by clobbering PE headers,
is the `EfiMain` from `StandaloneMmCoreEntryPoint.h`. This function does
not return any status code. In the case of `MmIplX64Relay`, the entry
point was forced to be cast to `STANDALONE_MM_FOUNDATION_ENTRY_POINT` and
return was checked.

This change will not check the return status and set `ReturnStatus` to
`EFI_SUCCESS` after the MM core returns.