MmSupervisorPkg: MmDxeSupport: Remove check for GCC compiled module
MM supervisor embedded version number and patch level numbers into the
final PE image, by specifying these numbers through linker flags and
GenFw flags.

However, this is not applicable for modules built through GCC. Because
the original output is in ELF format, and ELF to PE conversion is handled
through GenFw. GenFw will not be able to pertain the intended optional
header information, which does not exist in the ELF binary and cannot be
set by the objcopy.

This change simply do not evaluate this information in the case of GCC
build.