MmSupervisorPkg/Test/MmSupvRequestUnitTestApp: Update Request Comm Buffer test logic
The test currently calls `MmSupvRequestDxeToMmCommunicate()` and
checks if the return status code (the result from the comm buffer
request - `MM_SUPERVISOR_REQUEST_HEADER *)CommHeader->Data)->Result`)
is an `EFI_ERROR`.

Because the value is read from the comm buffer result, it will be
`EFI_ACCESS_DENIED` which is expected for the test to pass.

This change removes that check (to not impact other
`MmSupvRequestDxeToMmCommunicate()` callers) and continues to check
if the status code is `EFI_ACCESS_DENIED`.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>