MmSupervisorPkg/Core: Remove redundant null check (CodeQL)
The main change is to remove a NULL check for the CommBufferSize
parameter value in the MM_SUPERVISOR_REQUEST_FETCH_POLICY switch
case since it triggers a CodeQL alert as a redundant NULL check
operation. The actual parameter value is checked for NULL at the
beginning of the function and not modified until that line of code.

However, the function API and implementation is confusing so that's
cleaned up as well.

1. The Context parameter is not used at all. Since it is marked
   optional, it is left so the function prototype can remain
   consistent in case the parameter is needed in the future.
2. The CommBuffer and CommBufferSize parameters are also marked
   optional, but they're not. The function immediately returns
   EFI_INVALID_PARAMETER if they are not provided (null).

For (2), the optional modifier is simply removed from the arguments
to indicate the function expects valid pointers are passed.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>