gen_aux: Prioritize Data Symbols during Symbol Conflict
Multiple Symbol types and symbol data can exist for the same symbol
in the pdb file. This is mainly between labels and data symbol types.
This change prioritizes the data symbol type over the label symbol and
also adds a check to ensure that if there are two data type symbols for
the same symbol, the non-primitive data type is selected.