Add key symbols to the auxillary file header
This commit adds a list of signature / offset pairs to the header of the
auxillary file. This will allows a consumer of the auxillary file to
define a set of key symbols that are necessary before performing any
validation of an efi. The expectation is that the signature will allow
the consumer to identify the purpose of the provided offset and act
accordingly. The expectation is not that the signature will match the
symbol name, but the purpose that the symbol is used for.

[empty line]
[empty line]
Signed-off-by: Joey Vagedes <joey.vagedes@gmail.com>