Fix Dsc offsets in MmiEntrySpam.nasmb (#284)
## Description

The values used for the various DSC offsets in MmiEntrySpam.nasmb were
based on the normal SmiEntry.nasm used in the MM Supervisor. Instead,
the needed to align with the STM file structure. This change changes the
offsets to correctly match the structure they're referencing when
pulling variable data.

Additionally, SmmStm.c was updated to reference mXdSupported instead of
gPatchDxSupported. This also included a check for setting
gPatchMsrIa32MiscEnableSupported based on the return value of
StandardSignatureIsAuthenticAMD().

Also 2 files in the spam core had uncrustified run on them to pass CI.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [ ] Impacts functionality?
- **Functionality** - Does the change ultimately impact how firmware
functions?
- Examples: Add a new library, publish a new PPI, update an algorithm,
...
- [ ] Impacts security?
- **Security** - Does the change have a direct security impact on an
application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [ ] Breaking change?
- **Breaking change** - Will anyone consuming this change experience a
break
    in build or boot behavior?
- Examples: Add a new library class, move a module to a different repo,
call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
- **Documentation** - Does the change contain explicit documentation
additions
    outside direct code modifications (and comments)?
- Examples: Update readme file, add feature readme file, link to
documentation
    on an a separate Web page, ...

## How This Was Tested

Tested on Intel physical platform. The correct offsets are being used to
get the relevant CS/DS/SS values.

## Integration Instructions

N/A