Fix override from CodeQL and cherry-picks from latest basecore (#331)
## Description

This change updated the override validation tags. It also integrates the
cherry-pick commit from
https://github.com/microsoft/mu_basecore/commit/4d722ac91608d3c4fbe807b418b77f2a6e279117.

For details on how to complete to complete these options and their
meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on pipeline and passed override check. Also booted on
Q35 branch.

## Integration Instructions

The change needs to pair with basecore commit
https://github.com/microsoft/mu_basecore/commit/4d722ac91608d3c4fbe807b418b77f2a6e279117
or later.