gen_rim: add signature structure generation capability  (#332)
## Description

Updates the `gen_rim` command line tool to now have two subcommands
-`Generate` and `Signing`.

`Generate` is the same tool as before, providing the ability to generate
a RIM for a STM binary. It adds new functionality to also be able to
update an existing RIM's `signature` value.

`Signing` provides new functionality that can generate
`Sig_structure<payload>` cbor binary. This binary is the portion of RIM
that is actually used as the digest when signing the RIM as a part of
the `CoseSign1` structure. The `Sig_structure<payload>` is defined in
[rfc9052](https://www.rfc-editor.org/rfc/rfc9052.html#name-signing-and-verification-pr).
This command can either take the same arguments as the `Generate`
function or consume an existing RIM.

### Expected Flows

#### Flow 1
1. Generate the RIM with `rim generate $(STM_PATH) -o $(RIM_PATH) ... `
2. Generate the Sig Structure with `rim signing $(RIM_PATH) --from-rim
... `
3. Produce the signature
4. Update the RIM with `rim generate $(RIM_PATH) --update-signature
--signature $(SIGNATURE)`

#### Flow 2
1. Generate the Sig Structure with `rim signing $(STM_PATH) ...`
2. Produce the signature
3. Generate the RIM with `rim generate $(STM_PATH) -o $(RIM_PATH)
--signature $(SIGNATURE) ...`

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested RIM generation still works as expected.
Tested Sig Structure is successfully generated.

## Integration Instructions

N/A