gen_aux: Allow validation rules to only apply to certain build targets
This change allows validation rules to only apply to certain build
targets (DEBUG, RELEASE, NOOPT). By default, validation rules apply to
all three, but a rule can be marked to only apply to a subset of them
by adding a `target` field to the rule, which is a list of the supported
targets.