MmSupervisorPkg: Add library to publish MM Supervisor version
Adds a new NULL class library called "DxeMmSupervisorVersionPublicationLib"
that can be linked to any DXE module on a platform to publish MM Supervisor
Version information. The library is optional for a platform to include as
it does not impact MM Supervisor functionality. However, firmware and OS
code may make use of the version information produced by this library if
the information is published.

The library is meant to be simple to integrate not introducing any new
dependency requirements on the module that it is linked against. It
abstracts what version information is currently published and the
implementation may change over time.

The MM Supevisor version information is queried and published as a UEFI
variable that is available in both the firmware boot and OS environment.

The variable information is in Guid/MmSupervisorVersion.h. Variable policy
is applied to the variable by this library as well.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>