[CHERRRY-PICK] MmSupervisorPkg: Add library to publish MM Supervisor version (#371)
Adds a new NULL class library called
"DxeMmSupervisorVersionPublicationLib" that can be linked to any DXE
module on a platform to publish MM Supervisor Version information. The
library is optional for a platform to include as it does not impact MM
Supervisor functionality. However, firmware and OS code may make use of
the version information produced by this library if the information is
published.

The library is meant to be simple to integrate not introducing any new
dependency requirements on the module that it is linked against. It
abstracts what version information is currently published and the
implementation may change over time.

The MM Supevisor version information is queried and published as a UEFI
variable that is available in both the firmware boot and OS environment.

The variable information is in Guid/MmSupervisorVersion.h. Variable
policy is applied to the variable by this library as well.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

- Verified debug prints during boot.
- Checked variable policy.
- Read UEFI variable in OS.
- Read the MM Supervisor version reported in the OS (version here was
tested against older MM Supervisor v11.0.1 release):

![scpc_verifier_mm_supv_ver](https://github.com/user-attachments/assets/e45f2b89-bb5c-4cb2-9558-46e05c28e29b)

Platforms using the MM Supevisor can optionally add

`NULL|MmSupervisorPkg/Library/DxeMmSupervisorVersionPublicationLib/DxeMmSupervisorVersionPublicationLib.inf`
to a platform DXE module to publish MM Supervisor version information.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>
(cherry picked from main branch commit e885259392329d19dc05f233d989480e7f525d16)