Separate .bss section from the .data sections in the RELEASE image (#386)
## Description

Separate the .bss section of the image in RELEASE builds. This is
because we typically merge all of the different .data sections in
RELEASE builds which normally includes the .bss section. This breaks SEA
validation as it expects the .bss section to be ignored during
validation but instead we're including it in the normal .data section
validation.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on a SEA physical platform. Validation completes successfully
with the change.

## Integration Instructions

N/A