Change override tag to tracking tag (#388)
## Description

Changed the override tag for StandaloneMmCore.inf to a tracking tag and
we now also track the release branches. This is because some consumers
of the supervisor only use the release/branches instead of the dev
branch.


- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on physical platform. The platform can be built while using the
release branches with this change.

## Integration Instructions

N/A