Make it so PiSmmCpuSmiEntryFixupAddress is only called when not using SEA (#396)
## Description

Currently we touch up the SmiEntry.nasm file when calling
PiSmmCpuSmiEntryFixupAddress which is something we don't want to do when
using the SEA MmiEntrySea.nasmb file. This change makes it so
PiSmmCpuSmiEntryFixupAddress is only called when we aren't using a
different MmiEntry._>

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on physical system. The SmiEntry.nasm is no longer being fixed
up.

## Integration Instructions

N/A