Switch target filtering to scopes (#397)
## Description

Previously, each rule could apply to a subset of 3 different targets
(release, debug, noopt) and during auxiliary file generation, a single
target was passed to filter the applied rules.

To extend this functionality, this logic is now reversed, and scopes can
be any string, instead of limited to release, debug, and noopt:

1. Previously, a platform passed a single target and a rule could
support multiple targets
2. Now, A platform passes multiple scopes, and a rule only supports one
scope
3. If a rule does not specify a scope, it is always applied

- [ ] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Confirmed the same config file is generated based off of a scope similar
to the previous target

## Integration Instructions

Platforms must update any `[[rule]]` that has a `target = [...]` section
to now be `scope = ""`. If a `[[rule]]` had multiple `target` values,
the platform will need to determine any appopriate scope updates or rule
additions to account for that, as each `[[rule]]` now only has one
scope.