gen_aux: Add support for requiring all symbols have a rule
Adds a configuration file config option no_missing_rules that when set,
will cause the tool to exit with an error if any symbol does not have a
rule in the configuration file. Users can use `exclude_symbols` to
specify symbols that should be excluded from this check.

With the addition of other configs, we also move these config values to
a `[config]` section in the configuration file.