gen_aux: Add support for requiring all symbols have a rule (#394)
## Description

Adds a configuration file config option `no_missing_rules` that when
set, will cause the tool to exit with an error if any symbol does not
have a rule in the configuration file. Any `Procedure` symbols are
ignored in this requirement. Any symbols in `excluded_symbols` are
ignored in this requirement.

Adds a config option `excluded_symbols` that when set, will remove the
symbols from list of parsed pdb symbols. This effects both
`no_missing_rules` and `autogen`.


- [ ] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated the "status quo" did not change, but that setting the flag to
true results in a build break that prints the symbol and its address.

## Integration Instructions

Consumers may set the `no_missing_rules` to true to enable this
functionality.

Consumers must add a `[config]` section header above any currently
existing `autogen` configuration

e.g.

``` toml
autogen = false
```

to

``` toml
[config]
autogen = false
```

---------

Signed-off-by: Joey Vagedes <joey.vagedes@gmail.com>
Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>