[REBASE&FF] image_validation: bugfix aux file bounds checks (#401)
## Description

This commit fixes two separate issues with the bounds checks in the
image validation code that ensures certain pieces of data in the
auxiliary file are within the bounds of the file, according to the size
expectations described in the validation entry headers.

1. The first issue resolved is specific to the IMAGE_VALIDATION_CONTENT
validation type. In this validation type, the header contains N number
of additional bytes, that are the expected value that the particular
content should be. The purpose of this validation is to ensure that the
content provided by the auxiliary file for validation does not extend
beyond the bounds of the file.

The issue was that the bounds check only validated that the overall size
of the header + content was less than the overall size of the auxiliary
file. This has been updated to ensure that:

addr_of(header) + size_of(header) + size_of(content) <= addr_of(aux) +
size_of(aux)

2. The second is that only one validation type had a bounds check to
ensure the default value that we revert data to was within the bounds of
the auxiliary file. This has been moved out of the specific validation
type and is now a general check immediately before the reversion of the
particular data.

- [ ] Impacts functionality?
- [x] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

TestApp Continues to pass

## Integration Instructions

N/A