Move the GDT data to be read only data section (#381)
## Description

This change moves the GDT data to be read only section.

Note that `TssDescriptor` symbol is also going to be manipulated. But
execution-wise, this is always updated after disabling data writing. SEA
validation-wise, this change will enforce a rule of `TssDescriptor` to
be added for better validation.
 
- [x] Impacts functionality?
- [x] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change was tested on QEMU Q35 and booted to UEFI shell.

## Integration Instructions

N/A