GenSeaArtifacts: Simplify Integration to use
Update the cargo run commands to be called from the directory of the
tool itself. This simplifies the integration of these executables as
consumers no longer need to set up a virtual workspace at the root of
their repository to run the tools.

Each package is now its own virtual workspace, so that these tools do
not need to be added to the consumer's workspace. The consumers cargo
configuration files (.cargo/config.toml) will still control
configuration of the tools (for things like registry overrides and such)
due to how cargo works.