auxgen: Ignore labels for required symbols config (#417)
## Description

Ignores Label symbol for the "no_missing_rules" configuration value, as
labels are just addresses in the code section. Not only are they a
non-writeable section, but they are additionally not variables that can
be manipulated. They are simply addresses in the code section.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated Label Symbols are no longer required if said symbol is missing
a rule

## Integration Instructions

N/A - However users can remove labels from the excluded symbols path, or
their rules if they wish.