gen_aux: raw data report updates (#425)
## Description

Updates the raw data report with the following:

1. A new top level entry, sections, that contains a list of the PDB
sections that are present in the PDB (and a single non-section section
representing the PE Header).

2. Updates segments representing a symbol with the validation type of
`ValidationType::None` to be considered covered rather than uncovered
with the understanding that this covered reason is extremely weak.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated the json raw data output now contains the following:

```json
"sections": [
    {
      "name": "PE Header",
      "start": "0x0",
      "end": "0x1000"
    },
    {
      "name": ".text",
      "start": "0x1000",
      "end": "0x28000"
    },
    {
      "name": ".rdata",
      "start": "0x28000",
      "end": "0x37000"
    },
   ...
]
```

Validated that "Validation Rule: None" types are now marked as covered

```json
    {
      "symbol": "gMmCoreMmst.CurrentlyExecutingCpu",
      "start": "0x37388",
      "end": "0x37390",
      "covered": true,
      "reason": "Validation Rule: None"
    },
```
## Integration Instructions

N/A