Fixed issues with the pointer validation code (#428)
## Description

The pointer validation entry type has some issues with the validation
logic.
1. It doesn't correctly increment with the entry list iteration leading
to us point to nonvalid memory.
2. The calculation for the MSEG region was incorrect and was providing a
random address for BASE and the incorrect number of CPU cores.
3. The target address was incorrectly being set to the location of
symbol in the MM Supervisor image. It has been updated to use the
address of pointer so we can correctly judge if it's in or outside MSEG.

This changes address both of these issues by casting the entry type when
incrementing to address (1) and then using the correct data structures
to calculate the MSEG region.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on Intel physical platform. The validation process correctly gets
MSEG.

## Integration Instructions

N/A