Fix MSEG boundary check (#446)
## Description

The boundary check fix was lost between the testing branch and PR
branch. This PR changes the boundary check to be the correct range and
so we use the correct target address.

Additionally, there was a bug with the validation for some rules for
pointers, memory attributes and self reference. Before validating any of
the config file rules we relocate the target image through the function
`PeCoffLoaderRevertRelocateImage()`. To validate the original content of
the rules we now look at the original image instead of the target image
and update the target image when rules are successfully validated. This
maintains the expected behavior of validation process and also supports
multiple rules for the same symbol if that functionality is added in the
future.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on Intel Physical platform. The range and target address were
correct.

## Integration Instructions

N/A