Remove/Modify some global variables (#452)
## Description

Some global variables exist that are unnecessary because they're values
are constant or they're only used in testing scenarios. This change
removes the unnecessary variables and adds guards around the globals
that are used for testing so they only are in builds which enable the
testing PCD.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on Physical platform. Behavior remains the same with or without
this change.

## Integration Instructions

N/A