aux_gen: Auto generate rules for padding (#447)
## Description

While developing the coverage report raw data, it was noted that some
spaces between symbols in the .data section were not associated with any
symbols in the PDB. It was determined that this was padding in the
symbols themselves which work to align them in the file. This pull
request does the following things:

1. Cleans up the final build into multiple separate functions
2. Generates a content validation rule (of all zero) for any byte ranges
not assocaited with a symbol
3. Generates a content validation rule (of all zeros) for any padding
between sections, that is in a section that is both Read and Write.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

<_Describe the test(s) that were run to verify the changes._>

## Integration Instructions

<_Describe how these changes should be integrated. Use N/A if nothing is
required._>