gen_aux: Deny unknown config file fields (#456)
## Description

Disables the ability to put unexpected fields in top level fields the
following tables: `[[KeySymbol]]`, `[[rule]]`, `[config]`, and the top
level table of the file itself.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated that the file is properly parsed as is, and that putting
miscellaneous data in these tables causes the parser to fail.

## Integration Instructions

N/A