aux_gen: properly calculate array length
A bug was introduced in the code that calculates the total element count
of an array because it did not consider the case for when the symbol is
a multi-dimensional array. This is because the `total_size` function
returns the total size of the element in the array, which when working
with a multi-dimensional array, is another array, instead of the
underlying element.

To resolve this issue, we instead calculate the total element count by
using the `element_size()` function, which returns the size of the
underlying type instead. Due to this recursive nature of
`TypeInfo::from_type_index`, this is ultimately the actual element type,
even if the symbol is a 2, 3, 4, etc. dimensional array.