Update self ref validation to use the original image for validation (#461)
## Description

Update self ref validation to use the original image for validation.
This avoids the issue where the relocation code changes the target image
contents that we're trying to validate leading to a failure in the rules
validation. Instead we use the original image to validate the pointer.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on an Intel Physical platform. The previously failing rules now
pass.

## Integration Instructions

N/A