PeCoffValidationLib: Add host validation support
Adds host-based validation support to the `PeCoffImageValidationSelRef`
function by comparing the two value's RVAs instead of exact addresses.

This is necessary for local validation due to the fact that the the
self reference validation rule is simply validating that the address
held by a pointer is equal to some other address (typically a function
pointer) in the same image. This is done by grabbing the actual value
of the pointer and ensuring that it is equal to the `base_address` +
`offset` as specified in the auxiliary file. This works fine when
validating on the host, but when validating locally, the `base address`
is now a random address, but the true pointer value in the dumped image
remains the same.

To account for this, we simply compare the RVAs (relative to the
original image base address) of the two pointers instead of their actual
addresses.