Move core private data to local variable and  (#488)
## Description

This pull request moves the core private data to a local variable and
updates the code to use the local variable instead of the global
variable.

Additionally, this change moved the
`DiscoverStandaloneMmDriversInFvHobs` to after memory service is ready.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested with QEMU Q35 + 2502 basecore, and system boot to UEFI shell.

## Integration Instructions

N/A