MmSupervisorPkg: Add missing MmCoreDataHobGuid. (#490)
## Description

MmSupervisorCore relies on the gMmCoreDataHobGuid, but it is not
explicit in the need for it in the INF.

Add the gMmCoreDataHobGuid to explicitly define the dependency.

Fixes the GUID being drop in
https://github.com/microsoft/mu_feature_mm_supv/commit/0a59afaeb812d8241a0a760d4aed87ddd289c103.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Found while integrating into a new platform. Build failed due to
unresolved external,
but build passed after adding guid to dependency.

## Integration Instructions

No integration necessary.