test-aux: Add scope parameter for aux generation (#493)
## Description

Filters the generated rules to only those rules without a scope or with
the specified scope. Previously, all rules were generated which would
cause test-aux to fail for symbols that did not actually exist.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated that we continue to compile and rules are filtered when a
scope exists that is not matched.

## Integration Instructions

N/A